Create Skyscrappers.java
find the largest square area that can fit within a cityscape represented by an array cityLine, where each element denotes the height of a skyscraper with a width of 1. The skyscrapers are adjacent to each other without gaps.

Input:

An array cityLine representing the heights of the skyscrapers. For example, cityLine = [1, 2, 3, 2, 1].

Output:

The largest square area that can fit within the cityscape. For the example input, the output is 4 (a 2x2 square).

The example provided illustrates the problem:

Code



cityLine = [1, 2, 3, 2, 1]

In this case, several 2x2 squares can fit within the configuration of skyscrapers, but no larger square can. Therefore, the largest square area is 2 * 2 = 4..